refactor(commands): extract /context and /vim commands into display.rs
