+++ src-rust/crates/commands/src/lib.rs
@@ -248,6 +248,8 @@ mod search;
 pub use search::*;
 mod session_tools;
 pub use session_tools::*;
+mod display;
+pub use display::*;
 mod ui_settings;
 use ui_settings::*;
 
@@ -273,8 +275,6 @@ pub struct ImportConfigCommand;
 pub struct ThinkingCommand;
 // New commands
 // Batch-1 new commands
-pub struct ContextCommand;
-pub struct VimCommand;
 pub struct VoiceCommand;
 pub struct UpgradeCommand;
 pub struct ReleaseNotesCommand;
@@ -1201,123 +1201,6 @@ impl SlashCommand for ThinkingCommand {
                  'think carefully before answering'.",
                 model
             ))
-        }
-    }
-}
-
-// ---- /context ------------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for ContextCommand {
-    fn name(&self) -> &str { "context" }
-    fn description(&self) -> &str { "Show context window usage (tokens used / available)" }
-    fn help(&self) -> &str {
-        "Usage: /context\n\n\
-         Displays the current context window utilization:\n\
-         - Estimated tokens consumed by current conversation\n\
-         - Context window limit for the active model\n\
-         - Percentage used"
-    }
-
-    async fn execute(&self, _args: &str, ctx: &mut CommandContext) -> CommandResult {
-        let model = ctx.config.effective_model();
-
-        // Determine context window size from known model names
-        let context_window: u64 = if model.contains("claude-3-5") || model.contains("claude-3.5") {
-            200_000
-        } else if model.contains("opus") {
-            200_000
-        } else if model.contains("sonnet") {
-            200_000
-        } else if model.contains("haiku") {
-            200_000
-        } else {
-            200_000 // safe default for any Claude model
-        };
-
-        let used_tokens = ctx.cost_tracker.total_tokens();
-        let pct = if context_window > 0 {
-            (used_tokens as f64 / context_window as f64) * 100.0
-        } else {
-            0.0
-        };
-
-        let bar_width = 40usize;
-        let filled = ((pct / 100.0) * bar_width as f64).round() as usize;
-        let bar: String = "█".repeat(filled) + &"░".repeat(bar_width.saturating_sub(filled));
-
-        // Estimate approximate message tokens from the message list
-        let msg_char_count: usize = ctx.messages.iter().map(|m| m.get_all_text().len()).sum();
-        // Rough estimate: ~4 chars per token for message text
-        let msg_token_estimate = msg_char_count / 4;
-
-        CommandResult::Message(format!(
-            "Context Window Usage\n\
-             ────────────────────\n\
-             Model:          {model}\n\
-             Context window: {window:>10} tokens\n\
-             API tokens used:{used:>10} tokens  ({pct:.1}%)\n\
-             Est. msg size:  {msg:>10} tokens  (approx)\n\
-             Messages:       {msgs:>10}\n\n\
-             [{bar}] {pct:.1}%\n\n\
-             Use /compact to reduce context usage.",
-            model = model,
-            window = context_window,
-            used = used_tokens,
-            pct = pct,
-            msg = msg_token_estimate,
-            msgs = ctx.messages.len(),
-            bar = bar,
-        ))
-    }
-}
-
-// ---- /vim (/vi) ----------------------------------------------------------
-
-#[async_trait]
-impl SlashCommand for VimCommand {
-    fn name(&self) -> &str { "vim" }
-    fn aliases(&self) -> Vec<&str> { vec!["vi"] }
-    fn description(&self) -> &str { "Toggle vim keybinding mode on/off" }
-    fn help(&self) -> &str {
-        "Usage: /vim [on|off]\n\n\
-         Toggles vim keybinding mode in the REPL input.\n\
-         When enabled, use Esc to switch between INSERT and NORMAL modes.\n\n\
-         The setting is persisted to ~/.claurst/ui-settings.json."
-    }
-
-    async fn execute(&self, args: &str, _ctx: &mut CommandContext) -> CommandResult {
-        let current = load_ui_settings();
-        let current_mode = current.editor_mode.as_deref().unwrap_or("normal");
-
-        let new_mode = match args.trim() {
-            "on" | "vim" => "vim",
-            "off" | "normal" => "normal",
-            "" => {
-                // Toggle
-                if current_mode == "vim" { "normal" } else { "vim" }
-            }
-            other => {
-                return CommandResult::Error(format!(
-                    "Unknown argument '{}'. Use: /vim [on|off]",
-                    other
-                ))
-            }
-        };
-
-        match mutate_ui_settings(|s| s.editor_mode = Some(new_mode.to_string())) {
-            Ok(_) => CommandResult::Message(format!(
-                "Editor mode set to {}.\n{}",
-                new_mode,
-                if new_mode == "vim" {
-                    "Use Esc to switch between INSERT and NORMAL modes.\n\
-                     Restart the REPL for the change to take effect."
-                } else {
-                    "Using standard (readline-style) keyboard bindings.\n\
-                     Restart the REPL for the change to take effect."
-                }
-            )),
-            Err(e) => CommandResult::Error(format!("Failed to save setting: {}", e)),
         }
     }
 }
